Poprawka zadania suma liczb

+++ zajecia/Michalisko/sumaliczb.cpp
@@ -2,13 +2,13 @@
 using namespace std;
 
 int main(){
-int b;
-for (int a=0; a<1000000000; a++){
-if (a%2==0){
-} 
-else {
-b=a+b;
-}
-}
-cout << b <<endl;
+	long long int b;
+	for (long long int a=0; a<1000000000; a++)
+	{
+		if (a%2==0){
+		} else {
+			b=a+b;
+		}
+	}
+	cout << b <<endl;
 }
